feat: 삭제 예외 처리 구현

+++ src/main/java/com/likelion/wisesaying/service/SayingService.java
@@ -25,6 +25,11 @@ public class SayingService {
 
     public Long delete(String request) {
         Long deleteId = converter.splitId(request);
+
+        if (repository.findOne(deleteId) == null) {
+            throw new IllegalArgumentException(deleteId + "번 명언은 존재하지 않습니다.");
+        }
+
         repository.delete(deleteId);
         return deleteId;
     }
